add default discovery selection. either with a 'default' config or just being a singleton.
