+++ test/cli-test.js
@@ -1,0 +1,20 @@
+
+var Logger = require('../lib/logging/logger');
+
+require('mocha-sinon');
+var sinon = require("sinon");
+var assert = require('assert');
+var url = require('url');
+var _ = require('underscore');
+
+describe("CLI Behaviour", function () {
+
+    beforeEach(function() {
+
+    });
+
+    it("", function () {
+
+    });
+
+});
